pyproject.toml
[tool.poetry]
name = "grpc-vacancy-load-test"
version = "0.1.0"
description = ""
authors = ["Andrea Evangelista <andrea.itsec.eu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
locust = "^2.29.0"
grpcio = "^1.64.1"
grpcio-tools = "^1.64.1"
pytest = "^8.2.2"
pydantic = {extras = ["email"], version = "^2.7.4"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
